Move relevant circt-lec files to CIRCT library
