+++ cmake/modules/AddCIRCT.cmake
@@ -59,3 +59,8 @@ function(add_circt_translation_library name)
   set_property(GLOBAL APPEND PROPERTY CIRCT_TRANSLATION_LIBS ${name})
   add_circt_library(${ARGV} DEPENDS circt-headers)
 endfunction()
+
+function(add_circt_verification_library name)
+  set_property(GLOBAL APPEND PROPERTY CIRCT_VERIFICATION_LIBS ${name})
+  add_circt_library(${ARGV} DEPENDS circt-headers)
+endfunction()
