+++ lib/CMakeLists.txt
@@ -12,3 +12,4 @@ add_subdirectory(Scheduling)
 add_subdirectory(Support)
 add_subdirectory(Target)
 add_subdirectory(Transforms)
+add_subdirectory(Verification)
