+++ lib/Verification/Solver.cpp
@@ -10,10 +10,10 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#include "Solver.h"
-#include "Circuit.h"
-#include "LogicExporter.h"
-#include "Utility.h"
+#include "circt/Verification/Solver.h"
+#include "circt/Verification/Circuit.h"
+#include "circt/Verification/LogicExporter.h"
+#include "circt/Verification/Utility.h"
 #include "mlir/IR/Builders.h"
 #include <string>
 #include <z3++.h>
